requirements.txt
streamlit          1.35.0
google-generativeai 0.8.5
pandas             2.2.3
pillow             10.4.0